requirements.txt
paho-mqtt
docker
requests
python-logstash-async
psutil